🚚Moved pairs controller and services files one level up
